- updated README to add missing intrusction on linux's Makefile modification needed to built-in compilation
- updated Kconfig, switch RELAY & DBEUG_FS from forced selection to
dependancies

- added an atomic variable specify if lemona is done initializing
(useful only while compiled as a module to avoid race condition)

- TODO: there is a flaw in the blades right now, we've got no mean of
calculating and getting the content of buffers as it is. I need to
find a solution to that. Maybe finding a mean to pass more than one
param to a blade? But still, doesn't solve the filling of argsz and
extsz. Maybe this should be done by the blade done.